Refactored proxy configurat ion and updated documentat ion (#27) Renamed FileConfig Provider to FileProxyP rovider and updated its methods to handle proxy configurat ions Updated Proxy entity to include new fields and methods for conversion to envoy proto objects Updated SpringConf iguration to reflect changes in ProxyProvi der Updated DiscoveryS erver class comments for better understand ing Updated applicatio n.yml to include listener configurat ions Added details about envoy control plane and open API exemplar in README.md
