--- envoy/configs/clusters.yaml
@@ -1,18 +1,0 @@
-clusters:
-- name: service_envoyproxy_io
-  connectTimeout: 40
-  type: LOGICAL_DNS
-  lbPolicy: ROUND_ROBIN
-  hosts:
-  - socketAddress:
-      address: www.envoyproxy.io
-      port: 443
-
-- name: openapi-exemplar
-  connectTimeout: 30
-  type: LOGICAL_DNS
-  lbPolicy: ROUND_ROBIN
-  hosts:
-  - socketAddress:
-      address: openapi-exemplar.default.svc
-      port: 80
