--- envoy/configs/listeners.yaml
@@ -1,5 +1,0 @@
-listeners:
-- name: http
-  socketAddress:
-    address: 0.0.0.0
-    port: 10000
