--- envoy/configs/routes.yaml
@@ -1,14 +1,0 @@
-virtualHosts:
-- name: chained_envoy_hosts
-  domains: ["envoy.local"]
-  routes:
-  - match:
-      prefix: "/exemplar"
-    cluster: openapi-exemplar
-    mutations:
-      prefixRewrite: "/"
-  - match:
-      prefix: "/"
-    cluster: service_envoyproxy_io
-    mutations:
-      prefixRewrite: "/"
