+++ envoy/src/main/kotlin/ayansen/playground/envoy/entity/ProxyProviderConfigurations.kt
@@ -1,0 +1,8 @@
+package ayansen.playground.envoy
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "config-provider")
+data class ProxyProviderConfigurations(var file: FileProviderConfiguration = FileProviderConfiguration()) {
+    data class FileProviderConfiguration(var path: String = "")
+}
