+++ envoy/src/main/resources/application.yml
@@ -44,3 +44,8 @@ config-provider:
   file:
     path: "./configs"
 
+listeners:
+  - name: http
+    socketAddress:
+      address: 0.0.0.0
+      port: 10000
